Replace PrintResponse method
