+++ internal/handler/static/middleware.go
@@ -9,6 +9,7 @@ import (
 	"github.com/evg4b/uncors/internal/contracts"
 	"github.com/evg4b/uncors/internal/helpers"
 	"github.com/evg4b/uncors/internal/infra"
+	"github.com/evg4b/uncors/internal/tui"
 	"github.com/spf13/afero"
 )
 
@@ -42,7 +43,7 @@ func (h *Middleware) Wrap(next contracts.Handler) contracts.Handler {
 		}
 
 		http.ServeContent(response, request, stat.Name(), stat.ModTime(), file)
-		h.logger.PrintResponse(request, response.StatusCode())
+		tui.PrintResponse(request, response.StatusCode())
 	})
 }
