+++ internal/log/logger_test.go
@@ -3,11 +3,8 @@ package log_test
 
 import (
 	"bytes"
-	"net/http"
-	"net/url"
 	"testing"
 
-	"github.com/evg4b/uncors/internal/contracts"
 	"github.com/evg4b/uncors/internal/log"
 	"github.com/evg4b/uncors/testing/testutils"
 	"github.com/pterm/pterm"
@@ -104,118 +101,4 @@ func TestPrefixedLogger(t *testing.T) {
 
 		assert.Equal(t, expected, output.String())
 	}))
-
-	t.Run("printResponse", func(t *testing.T) {
-		log.EnableColor()
-		logger := log.NewLogger(prefix, log.WithStyle(&pterm.Style{pterm.FgBlack, pterm.BgLightBlue}))
-
-		t.Run("should correctly format", testutils.LogTest(func(t *testing.T, output *bytes.Buffer) {
-			tests := []struct {
-				name       string
-				response   *http.Response
-				request    *contracts.Request
-				statusCode int
-				expected   string
-			}{
-				{
-					name:       "response with status code 1xx",
-					statusCode: 100,
-					request:    request(http.MethodPost, "/api/info"),
-					expected: "\x1b[30;104m\x1b[30;104m  Test  \x1b[0m\x1b[0m \x1b[39;49m\x1b[39;49m\x1b[30;46m" +
-						"\x1b[30;46m 100 POST \x1b[0m\x1b[39;49m\x1b[0m\x1b[39;49m \x1b[96m\x1b[96mhttps://api.domain" +
-						".com/api/info\x1b[0m\x1b[39;49m\x1b[0m\x1b[39;49m\x1b[0m\x1b[0m\n",
-				},
-				{
-					name:       "response with success code 2xx",
-					statusCode: 200,
-					request:    request(http.MethodGet, "/help"),
-					expected: "\x1b[30;104m\x1b[30;104m  Test  \x1b[0m\x1b[0m \x1b[39;49m\x1b[39;49m\x1b[30;46m" +
-						"\x1b[30;46m 100 POST \x1b[0m\x1b[39;49m\x1b[0m\x1b[39;49m \x1b[96m\x1b[96mhttps://api." +
-						"domain.com/api/info\x1b[0m\x1b[39;49m\x1b[0m\x1b[39;49m\x1b[0m\x1b[0m\n\x1b[30;104m\x1b" +
-						"[30;104m  Test  \x1b[0m\x1b[0m \x1b[39;49m\x1b[39;49m\x1b[30;42m\x1b[30;42m 200 GET \x1b" +
-						"[0m\x1b[39;49m\x1b[0m\x1b[39;49m \x1b[32m\x1b[32mhttps://api.domain.com/help\x1b[0m\x1b" +
-						"[39;49m\x1b[0m\x1b[39;49m\x1b[0m\x1b[0m\n",
-				},
-				{
-					name:       "response with redirect code 3xx",
-					statusCode: 300,
-					request:    request(http.MethodPatch, "/api/user"),
-					expected: "\x1b[30;104m\x1b[30;104m  Test  \x1b[0m\x1b[0m \x1b[39;49m\x1b[39;49m\x1b[30;46m\x1b" +
-						"[30;46m 100 POST \x1b[0m\x1b[39;49m\x1b[0m\x1b[39;49m \x1b[96m\x1b[96mhttps://api.domain" +
-						".com/api/info\x1b[0m\x1b[39;49m\x1b[0m\x1b[39;49m\x1b[0m\x1b[0m\n\x1b[30;104m\x1b[30;104m  " +
-						"Test  \x1b[0m\x1b[0m \x1b[39;49m\x1b[39;49m\x1b[30;42m\x1b[30;42m 200 GET \x1b[0m\x1b[39;" +
-						"49m\x1b[0m\x1b[39;49m \x1b[32m\x1b[32mhttps://api.domain.com/help\x1b[0m\x1b[39;49m\x1b[0m" +
-						"\x1b[39;49m\x1b[0m\x1b[0m\n\x1b[30;104m\x1b[30;104m  Test  \x1b[0m\x1b[0m \x1b[39;49m\x1b" +
-						"[39;49m\x1b[30;43m\x1b[30;43m 300 PATCH \x1b[0m\x1b[39;49m\x1b[0m\x1b[39;49m \x1b[33m\x1b" +
-						"[33mhttps://api.domain.com/api/user\x1b[0m\x1b[39;49m\x1b[0m\x1b[39;49m\x1b[0m\x1b[0m\n",
-				},
-				{
-					name:       "response with user request error code 4xx",
-					statusCode: 400,
-					request:    request(http.MethodDelete, "/api/user/permission"),
-					expected: "\x1b[30;104m\x1b[30;104m  Test  \x1b[0m\x1b[0m \x1b[39;49m\x1b[39;49m\x1b[30;46m" +
-						"\x1b[30;46m 100 POST \x1b[0m\x1b[39;49m\x1b[0m\x1b[39;49m \x1b[96m\x1b[96mhttps://api." +
-						"domain.com/api/info\x1b[0m\x1b[39;49m\x1b[0m\x1b[39;49m\x1b[0m\x1b[0m\n\x1b[30;104m\x1b" +
-						"[30;104m  Test  \x1b[0m\x1b[0m \x1b[39;49m\x1b[39;49m\x1b[30;42m\x1b[30;42m 200 GET \x1b" +
-						"[0m\x1b[39;49m\x1b[0m\x1b[39;49m \x1b[32m\x1b[32mhttps://api.domain.com/help\x1b[0m\x1b" +
-						"[39;49m\x1b[0m\x1b[39;49m\x1b[0m\x1b[0m\n\x1b[30;104m\x1b[30;104m  Test  \x1b[0m\x1b[0m \x1b" +
-						"[39;49m\x1b[39;49m\x1b[30;43m\x1b[30;43m 300 PATCH \x1b[0m\x1b[39;49m\x1b[0m\x1b[39;49m " +
-						"\x1b[33m\x1b[33mhttps://api.domain.com/api/user\x1b[0m\x1b[39;49m\x1b[0m\x1b[39;49m\x1b" +
-						"[0m\x1b[0m\n\x1b[30;104m\x1b[30;104m  Test  \x1b[0m\x1b[0m \x1b[39;49m\x1b[39;49m\x1b" +
-						"[30;101m\x1b[30;101m 400 DELETE \x1b[0m\x1b[39;49m\x1b[0m\x1b[39;49m \x1b[91m\x1b" +
-						"[91mhttps://api.domain.com/api/user/permission\x1b[0m\x1b[39;49m\x1b[0m\x1b[39;49m" +
-						"\x1b[0m\x1b[0m\n",
-				},
-				{
-					name:       "response with internal server error code 5xx",
-					statusCode: 500,
-					request:    request(http.MethodPost, "/"),
-					expected: "\x1b[30;104m\x1b[30;104m  Test  \x1b[0m\x1b[0m \x1b[39;49m\x1b[39;49m\x1b[30;46m\x1b" +
-						"[30;46m 100 POST \x1b[0m\x1b[39;49m\x1b[0m\x1b[39;49m \x1b[96m\x1b[96mhttps://api.domain." +
-						"com/api/info\x1b[0m\x1b[39;49m\x1b[0m\x1b[39;49m\x1b[0m\x1b[0m\n\x1b[30;104m\x1b[30;104m" +
-						"  Test  \x1b[0m\x1b[0m \x1b[39;49m\x1b[39;49m\x1b[30;42m\x1b[30;42m 200 GET \x1b[0m" +
-						"\x1b[39;49m\x1b[0m\x1b[39;49m \x1b[32m\x1b[32mhttps://api.domain.com/help\x1b[0m\x1b[39;" +
-						"49m\x1b[0m\x1b[39;49m\x1b[0m\x1b[0m\n\x1b[30;104m\x1b[30;104m  Test  \x1b[0m\x1b[0m \x1b" +
-						"[39;49m\x1b[39;49m\x1b[30;43m\x1b[30;43m 300 PATCH \x1b[0m\x1b[39;49m\x1b[0m\x1b[39;49m " +
-						"\x1b[33m\x1b[33mhttps://api.domain.com/api/user\x1b[0m\x1b[39;49m\x1b[0m\x1b[39;49m\x1b" +
-						"[0m\x1b[0m\n\x1b[30;104m\x1b[30;104m  Test  \x1b[0m\x1b[0m \x1b[39;49m\x1b[39;49m\x1b" +
-						"[30;101m\x1b[30;101m 400 DELETE \x1b[0m\x1b[39;49m\x1b[0m\x1b[39;49m \x1b[91m\x1b" +
-						"[91mhttps://api.domain.com/api/user/permission\x1b[0m\x1b[39;49m\x1b[0m\x1b[39;49m\x1b" +
-						"[0m\x1b[0m\n\x1b[30;104m\x1b[30;104m  Test  \x1b[0m\x1b[0m \x1b[39;49m\x1b[39;49m\x1b" +
-						"[30;101m\x1b[30;101m 500 POST \x1b[0m\x1b[39;49m\x1b[0m\x1b[39;49m \x1b[91m\x1b" +
-						"[91mhttps://api.domain.com/\x1b[0m\x1b[39;49m\x1b[0m\x1b[39;49m\x1b[0m\x1b[0m\n",
-				},
-			}
-			for _, testCase := range tests {
-				t.Run(testCase.name, func(t *testing.T) {
-					logger.PrintResponse(testCase.request, testCase.statusCode)
-
-					assert.Equal(t, testCase.expected, output.String())
-				})
-			}
-		}))
-
-		t.Run("should panic for status code less then 100", testutils.LogTest(func(t *testing.T, _ *bytes.Buffer) {
-			assert.Panics(t, func() {
-				logger.PrintResponse(request(http.MethodGet, "/"), 50)
-			})
-		}))
-
-		t.Run("should panic for status code great then 599", testutils.LogTest(func(t *testing.T, _ *bytes.Buffer) {
-			assert.Panics(t, func() {
-				logger.PrintResponse(request(http.MethodGet, "/"), 600)
-			})
-		}))
-	})
-}
-
-func request(method string, path string) *http.Request {
-	return &http.Request{
-		Method: method,
-		URL: &url.URL{
-			Scheme: "https",
-			Host:   "api.domain.com",
-			Path:   path,
-		},
-	}
 }
